refactor: creating domain folder
