+++ src/common/domain/dtos/auth/make-login.dto.ts
@@ -1,0 +1,9 @@
+import { Field, InputType } from '@nestjs/graphql';
+
+@InputType()
+export class MakeLogin {
+  @Field({ nullable: false })
+  email: string;
+  @Field({ nullable: false })
+  password: string;
+}
