+++ src/common/domain/interfaces/logged-user.interface.ts
